Adds react-router and -dom dependencies, renders each page individually
