+++ client/Components/SwitchAndCheckout.jsx
