+++ client/index.js
@@ -1,11 +1,13 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import App from './App.js';
+import App from './App.jsx';
 import "./styles.css";
+import { BrowserRouter } from "react-router-dom";
+
 
 const root = createRoot(document.getElementById("root"));
 root.render(
-  <div className="main">
-  <App />
-  </div>
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
 );
