+++ package.json
@@ -18,7 +18,9 @@
     "express": "^4.18.2",
     "pg": "^8.11.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router": "^6.12.1",
+    "react-router-dom": "^6.12.1"
   },
   "devDependencies": {
     "@babel/core": "^7.22.5",
