+++ webpack.config.js
@@ -30,6 +30,7 @@ module.exports = {
     },
   output: {
     path: path.resolve(__dirname, 'build'),
+    publicPath: '/',
     filename: 'bundle.js',
   },
   plugins: [
@@ -43,6 +44,7 @@ module.exports = {
       directory: path.resolve(__dirname, 'build'),
       publicPath: '/',
     },
+    historyApiFallback: true,
     compress: true,
     port: 8080
   }
